Update debian/changelog if necessary to include dev release
If we're building a commit that is tagged with a release version that
matches debian/changelog, do nothing. Otherwise, add a new entry to
the top of the changelog for the dev release.

For example, if the latest version in the changelog is "1.2.4-1", and
the current commit is tagged "v1.2.4" do nothing. But if the current
commit is not tagged, add version "1.2.4+20130917230038.dev-1" to the
changelog.

This allows us to build normal releases as expected, as well as
development snapshot releases that superceed them.